added main flip and mirror controllers which control active matrix
